Consolidate onepassword unit tests so that ansible-test will find them when the plugin is modified.
