+++ tests/unit/plugins/lookup/onepassword_common.py
@@ -15,7 +15,7 @@ from ansible_collections.community.general.plugins.lookup.onepassword import (
 
 
 def load_file(file):
-    with open((os.path.join(os.path.dirname(__file__), "fixtures", file)), "r") as f:
+    with open((os.path.join(os.path.dirname(__file__), "onepassword_fixtures", file)), "r") as f:
         return json.loads(f.read())
