+++ tests/unit/plugins/lookup/onepassword_fixtures/v1_out_02.json.license
