+++ tests/unit/plugins/lookup/onepassword_fixtures/v1_out_03.json
