+++ tests/unit/plugins/lookup/onepassword_fixtures/v2_out_03.json
